% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{recode_sys_size}
\alias{recode_sys_size}
\title{Recode system size variable}
\usage{
recode_sys_size(data)
}
\arguments{
\item{data}{data.frame containing Gillespie data}
}
\value{
data.frame with system_size recoded
}
\description{
Recode system size variable
}
